+++ src/function/utils/import.js
@@ -1,12 +1,20 @@
 /**
  * Import functions from an object or a file
  * @param {function | String | Object} object
- * @param {boolean} [override]         If true, existing functions will be
- *                                     overwritten. False by default.
+ * @param {Object} [options]        Available options:
+ *                                  {Boolean} override
+ *                                  If true, existing functions will be
+ *                                  overwritten. False by default.
  */
 // TODO: return status information
-math['import'] = function math_import(object, override) {
+math['import'] = function math_import(object, options) {
     var name;
+    var opts = {
+        override: false
+    };
+    if (options && options instanceof Object) {
+        util.extend(opts, options);
+    }
 
     if (isString(object)) {
         // a string with a filename
@@ -23,7 +31,7 @@ math['import'] = function math_import(object, override) {
         // a single function
         name = object.name;
         if (name) {
-            if (override || math[name] === undefined) {
+            if (opts.override || math[name] === undefined) {
                 _import(name, object);
             }
         }
@@ -37,7 +45,7 @@ math['import'] = function math_import(object, override) {
             if (object.hasOwnProperty(name)) {
                 var value = object[name];
                 if (isSupportedType(value)) {
-                    if (override || math[name] === undefined) {
+                    if (opts.override || math[name] === undefined) {
                         _import(name, value);
                     }
                 }